Add stageDebuggable build type
